Remove most logic from root build.gradle file
