+++ buildSrc/src/main/kotlin/utils/Color.kt
@@ -1,0 +1,20 @@
+package utils
+
+internal enum class Color(ansiCode: Int) {
+    NONE(0),
+    BLACK(30),
+    RED(31),
+    GREEN(32),
+    YELLOW(33),
+    BLUE(34),
+    PURPLE(35),
+    CYAN(36),
+    WHITE(37);
+
+    private val ansiString: String = "\u001B[${ansiCode}m"
+
+    override fun toString(): String {
+        return ansiString
+    }
+}
+
