+++ buildSrc/src/main/kotlin/utils/string.kt
@@ -1,0 +1,5 @@
+package utils
+
+operator fun String.times(x: Int): String {
+    return List(x) { this }.joinToString("")
+}
