+++ faker/build.gradle.kts
@@ -4,24 +4,6 @@ import org.gradle.jvm.tasks.Jar
 plugins {
 }
 
-// no sources for this module
-sourceSets {
-    main.configure {
-        java { setSrcDirs(emptySet<String>()) }
-        kotlin { setSrcDirs(emptySet<String>()) }
-        resources { setSrcDirs(emptySet<String>()) }
-    }
-    test.configure {
-        java { setSrcDirs(emptySet<String>()) }
-        kotlin { setSrcDirs(emptySet<String>()) }
-        resources { setSrcDirs(emptySet<String>()) }
-    }
-}
-
-// disable api validation tasks
-tasks.apiBuild { enabled = false }
-tasks.apiCheck { enabled = false }
-tasks.apiDump { enabled = false }
 // disable the default jar task
 tasks.withType<Jar> { enabled = false }
 // never publish
