+++ gradle/libs.versions.toml
@@ -69,5 +69,7 @@ kotlinpoet = ["kotlinpoet", "kotlinpoet-ksp"]
 test-kotest = ["test-kotest-runner", "test-kotest-assertions-core", "test-kotest-property"]
 
 [plugins]
-kotlin-bcv = { id = "org.jetbrains.kotlinx.binary-compatibility-validator", version = "0.15.0-Beta.1" }
 # import plugins using Maven coordinates (see above), not the Gradle plugin ID
+kotlin-bcv = { id = "org.jetbrains.kotlinx.binary-compatibility-validator", version = "0.15.0-Beta.1" }
+nexus-publish = { id = "io.github.gradle-nexus.publish-plugin", version = "1.3.0" }
+benmanes-versions = { id = "com.github.ben-manes.versions", version = "0.51.0" }
